Extract YAML parsing and analysis into separate files
