+++ eslint-bridge/src/parser.ts
@@ -19,27 +19,12 @@
  */
 import * as fs from 'fs';
 import * as babel from '@babel/eslint-parser';
-import { AST, Linter, SourceCode } from 'eslint';
+import { Linter, SourceCode } from 'eslint';
 import * as VueJS from 'vue-eslint-parser';
 import * as tsEslintParser from '@typescript-eslint/parser';
 import { getContext } from './context';
 import { JsTsAnalysisInput, ParsingError } from './analyzer';
 import { getProgramById } from './programManager';
-// replace this with `yaml` when we deprecate node 12
-import * as yaml from 'yaml-node12';
-import { FileType, visit } from './utils';
-import { Comment, Node } from 'estree';
-import { assert } from 'console';
-
-type Lambda = {
-  code: string;
-  line: number;
-  column: number;
-  offset: number;
-  lineStarts: number[];
-  text: string;
-  format: 'PLAIN' | 'BLOCK_FOLDED' | 'BLOCK_LITERAL';
-};
 
 const babelParser = { parse: babel.parseForESLint, parser: '@babel/eslint-parser' };
 const vueParser = { parse: VueJS.parseForESLint, parser: 'vue-eslint-parser' };
@@ -225,246 +210,4 @@ function babelConfig(config: Linter.ParserOptions) {
     configFile: false,
   };
   return { ...config, requireConfigFile: false, babelOptions };
-}
-
-export function parseYaml(filePath: string): Lambda[] | ParsingError {
-  const src = getFileContent(filePath);
-  const lineCounter = new yaml.LineCounter();
-  const tokens = new yaml.Parser(lineCounter.addNewLine).parse(src);
-  // YAML supports a marker "---" that indicates the end of a document: a file may contain multiple documents
-  const docs = new yaml.Composer({ keepSourceTokens: true }).compose(tokens);
-
-  const lambdas: Lambda[] = [];
-  for (const doc of docs) {
-    // we only consider the first error
-    if (doc.errors.length > 0) {
-      const error = doc.errors[0];
-      return {
-        line: lineCounter.linePos(error.pos[0]).line,
-        message: error.message,
-        code: ParseExceptionCode.Parsing,
-      };
-    }
-
-    yaml.visit(doc, {
-      Pair(_, pair: any, ancestors: any) {
-        if (
-          (isInlineAwsLambda(pair, ancestors) || isInlineAwsServerless(pair, ancestors)) &&
-          isSupportedFormat(pair)
-        ) {
-          const { value, srcToken } = pair;
-          const code = srcToken.value.source;
-          // this should not happen
-          assert(code != null, 'An extracted inline JavaScript snippet should not be undefined.');
-          const [offsetStart] = value.range;
-          const { line, col: column } = lineCounter.linePos(offsetStart);
-          const lineStarts = lineCounter.lineStarts;
-
-          lambdas.push({
-            code,
-            line,
-            column,
-            offset: fixOffset(offsetStart, value.type),
-            lineStarts,
-            text: src,
-            format: pair.value.type,
-          });
-        }
-      },
-    });
-  }
-
-  function isInlineAwsLambda(pair: any, ancestors: any[]) {
-    return (
-      isZipFile(pair) &&
-      hasCode(ancestors) &&
-      hasNodeJsRuntime(ancestors) &&
-      hasType(ancestors, 'AWS::Lambda::Function')
-    );
-
-    function isZipFile(pair: any) {
-      return pair.key.value === 'ZipFile';
-    }
-    function hasCode(ancestors: any[], level = 2) {
-      return ancestors[ancestors.length - level]?.key?.value === 'Code';
-    }
-  }
-
-  function isInlineAwsServerless(pair: any, ancestors: any[]) {
-    return (
-      isInlineCode(pair) &&
-      hasNodeJsRuntime(ancestors, 1) &&
-      hasType(ancestors, 'AWS::Serverless::Function', 3)
-    );
-
-    // we need to check the pair directly instead of ancestors, otherwise it will validate all siblings
-    function isInlineCode(pair: any) {
-      return pair.key.value === 'InlineCode';
-    }
-  }
-
-  function isSupportedFormat(pair: yaml.Pair<any, any>) {
-    return ['PLAIN', 'BLOCK_FOLDED', 'BLOCK_LITERAL'].includes(pair.value?.type);
-  }
-
-  function hasNodeJsRuntime(ancestors: any[], level = 3) {
-    return ancestors[ancestors.length - level]?.items?.some(
-      (item: any) => item?.key.value === 'Runtime' && item?.value?.value.startsWith('nodejs'),
-    );
-  }
-
-  function hasType(ancestors: any[], value: string, level = 5) {
-    return ancestors[ancestors.length - level]?.items?.some(
-      (item: any) => item?.key.value === 'Type' && item?.value.value === value,
-    );
-  }
-
-  // the offset value needs to be fixed depending on the type of string format in YAML
-  function fixOffset(offset: number, format: string): number {
-    if (format === 'BLOCK_FOLDED' || format === 'BLOCK_LITERAL') {
-      return offset + 2;
-    } else {
-      return offset;
-    }
-  }
-
-  return lambdas;
-}
-
-/**
- * If there is at least 1 error in any JS lambda, we return only the first error
- */
-export function buildSourceCodesFromYaml(filePath: string): SourceCode[] | ParsingError {
-  const lambdasOrError = parseYaml(filePath);
-
-  const constainsError = !Array.isArray(lambdasOrError);
-  if (constainsError) {
-    return lambdasOrError;
-  }
-  const lambdas = lambdasOrError;
-
-  const sourceCodes: SourceCode[] = [];
-  for (const lambda of lambdas) {
-    const { code } = lambda;
-    /**
-     * The file path is left empty as it is ignored by `buildSourceCode` if the file content is provided, which
-     * happens to be the case here since we extract inline JavaScript code.
-     */
-    const input = { filePath: '', fileContent: code, fileType: FileType.MAIN, tsConfigs: [] };
-    const sourceCodeOrError = buildSourceCode(input, 'js');
-    if (sourceCodeOrError instanceof SourceCode) {
-      const patchedSourceCode = patchSourceCode(sourceCodeOrError, lambda);
-      sourceCodes.push(patchedSourceCode);
-    } else {
-      return patchParsingError(sourceCodeOrError, lambda);
-    }
-  }
-  return sourceCodes;
-
-  function patchSourceCode(originalSourceCode: SourceCode, lambda: Lambda) {
-    /* taken from eslint/lib/source-code/source-code.js#constructor */
-    function computeLines() {
-      const lineBreakPattern = /\r\n|[\r\n\u2028\u2029]/u;
-      const lineEndingPattern = new RegExp(lineBreakPattern.source, 'gu');
-      let match;
-      const lines = [];
-
-      let i = 0;
-      while ((match = lineEndingPattern.exec(lambda.text))) {
-        lines.push(lambda.text.slice(lambda.lineStarts[i], match.index));
-        i++;
-      }
-      lines.push(lambda.text.slice(lambda.lineStarts[lambda.lineStarts.length - 1]));
-
-      return lines;
-    }
-
-    /**
-     * Patches loc.start, loc.end and range in sourceCode.ast nodes
-     */
-    function patchLocations(sourceCode: SourceCode, lambda: Lambda) {
-      const { offset } = lambda;
-
-      visit(sourceCode, node => {
-        fixNodeLocation(node);
-      });
-
-      const { comments } = sourceCode.ast;
-      for (const comment of comments) {
-        fixNodeLocation(comment);
-      }
-
-      const { tokens } = sourceCode.ast;
-      for (const token of tokens) {
-        fixNodeLocation(token);
-      }
-
-      function fixNodeLocation(node: Node | Comment | AST.Token) {
-        if (node.loc != null && node.range != null) {
-          node.loc = {
-            start: sourceCode.getLocFromIndex(node.range[0] + offset),
-            end: sourceCode.getLocFromIndex(node.range[1] + offset),
-          };
-        }
-        if (node.range) {
-          const [sRange, eRange] = node.range;
-          node.range = [sRange + offset, eRange + offset];
-        }
-      }
-    }
-
-    /**
-     * 1. We compute lines here because sourceCode.lines is in the JS referrential and
-     * the YAML parser does not provide it
-     * 2. We override the values lineStartIndices, text and lines in sourceCode
-     * from the JS to the YAML referrential. We must use Object.create() because
-     * SourceCode's properties are frozen
-     * 3. We patch the sourceCode.ast nodes after 1 & 2 as it relies on values computed earlier
-     * 4. We rebuild sourceCode from the patched values because it builds internal properties that are dependent on them
-     */
-    const lines = computeLines();
-    const patchedSourceCode = Object.create(originalSourceCode, {
-      lineStartIndices: { value: lambda.lineStarts },
-      text: { value: lambda.text },
-      lines: { value: lines },
-    });
-
-    patchLocations(patchedSourceCode, lambda);
-
-    return new SourceCode({
-      text: patchedSourceCode.text,
-      ast: patchedSourceCode.ast,
-      parserServices: patchedSourceCode.parserServices,
-      scopeManager: patchedSourceCode.scopeManager,
-      visitorKeys: patchedSourceCode.visitorKeys,
-    });
-  }
-}
-
-function patchParsingError(parsingError: ParsingError, lambda: Lambda): ParsingError {
-  const { code, line, message } = parsingError;
-  let patchedLine: number | undefined;
-  let patchedMessage = message;
-  if (line) {
-    patchedLine = lambda.format === 'PLAIN' ? lambda.line : lambda.line + line;
-    patchedMessage = patchParsingErrorMessage(message, patchedLine, lambda);
-  }
-  return {
-    code,
-    line: patchedLine,
-    message: patchedMessage,
-  };
-}
-
-function patchParsingErrorMessage(message: string, patchedLine: number, lambda: Lambda) {
-  /* patching error message `(<line>:<column>)` */
-  const regex = /((?<line>\d+):(?<column>\d+))/;
-  const found = message.match(regex);
-  if (found) {
-    const line = found.groups!.line;
-    const column = Number(found.groups!.column);
-    const patchedColumn = lambda.format === 'PLAIN' ? column + lambda.column - 1 : column;
-    return message.replace(`(${line}:${column})`, `(${patchedLine}:${patchedColumn})`);
-  }
-  return message;
 }
