+++ eslint-bridge/tests/analyzer.test.ts
@@ -24,13 +24,11 @@ import {
   initLinter,
   analyzeCss,
   Issue,
-  analyzeYaml,
 } from 'analyzer';
 import { join } from 'path';
 import * as fs from 'fs';
 import { setContext } from 'context';
 import * as stylelint from 'stylelint';
-import { ParseExceptionCode } from '../src/parser';
 
 const noOneIterationIssue: Issue = {
   line: 3,
@@ -478,179 +476,5 @@ describe('#analyzeCss', () => {
   it('should throw when failed promise returned', async () => {
     (stylelint.lint as jest.Mock).mockRejectedValue(new Error('some reason'));
     await expect(analyzeCss(request)).rejects.toEqual(new Error('some reason'));
-  });
-});
-
-describe('#analyzeYaml', () => {
-  it('should analyze YAML file', async () => {
-    initLinter([
-      { key: 'no-all-duplicated-branches', configurations: [], fileTypeTarget: ['MAIN'] },
-    ]);
-    const {
-      issues: [issue],
-    } = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/valid-lambda.yaml'),
-      fileContent: undefined,
-      fileType: 'MAIN',
-      tsConfigs: [],
-    });
-    expect(issue).toEqual(
-      expect.objectContaining({
-        ruleId: 'no-all-duplicated-branches',
-        line: 8,
-        column: 17,
-        endLine: 8,
-        endColumn: 46,
-      }),
-    );
-  });
-
-  it('should return an empty issues list when parse error', async () => {
-    initLinter([
-      { key: 'no-all-duplicated-branches', configurations: [], fileTypeTarget: ['MAIN'] },
-    ]);
-    const { issues, parsingError } = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/invalid-yaml.yaml'),
-      fileContent: undefined,
-      tsConfigs: [],
-      fileType: 'MAIN',
-    });
-    expect(issues).toHaveLength(0);
-    expect(parsingError).toHaveProperty('code', ParseExceptionCode.Parsing);
-    expect(parsingError).toHaveProperty('line', 2);
-    expect(parsingError).toHaveProperty('message', 'Map keys must be unique');
-  });
-
-  it('should not break when using a rule with a quickfix', async () => {
-    initLinter([{ key: 'no-extra-semi', configurations: [], fileTypeTarget: ['MAIN'] }]);
-    const result = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/valid-serverless-quickfix.yaml'),
-      fileContent: undefined,
-      fileType: 'MAIN',
-      tsConfigs: [],
-    });
-    const {
-      issues: [
-        {
-          quickFixes: [quickFix],
-        },
-      ],
-    } = result;
-    expect(quickFix.edits).toEqual([
-      {
-        text: ';',
-        loc: {
-          line: 7,
-          column: 58,
-          endLine: 7,
-          endColumn: 60,
-        },
-      },
-    ]);
-  });
-
-  it('should not break when using "enforce-trailing-comma" rule', async () => {
-    initLinter([
-      {
-        key: 'enforce-trailing-comma',
-        configurations: ['always-multiline'],
-        fileTypeTarget: ['MAIN'],
-      },
-    ]);
-    const { issues } = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/fail-lambda.yaml'),
-      fileContent: undefined,
-      fileType: 'MAIN',
-      tsConfigs: [],
-    });
-    expect(issues).toHaveLength(2);
-    expect(issues[0]).toEqual(
-      expect.objectContaining({
-        line: 30,
-        column: 28,
-        endLine: 31,
-        endColumn: 0,
-      }),
-    );
-    expect(issues[1]).toEqual(
-      expect.objectContaining({
-        line: 31,
-        column: 19,
-        endLine: 32,
-        endColumn: 0,
-      }),
-    );
-  });
-
-  it('should not break when using a rule with secondary locations', async () => {
-    initLinter([{ key: 'no-new-symbol', configurations: [], fileTypeTarget: ['MAIN'] }]);
-    const result = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/valid-serverless-secondary.yaml'),
-      fileContent: undefined,
-      fileType: 'MAIN',
-      tsConfigs: [],
-    });
-    const {
-      issues: [
-        {
-          secondaryLocations: [secondaryLocation],
-        },
-      ],
-    } = result;
-    expect(secondaryLocation).toEqual({
-      line: 7,
-      column: 35,
-      endLine: 7,
-      endColumn: 41,
-    });
-  });
-
-  it('should not break when using a regex-parsing rule', async () => {
-    initLinter([{ key: 'sonar-no-regex-spaces', configurations: [], fileTypeTarget: ['MAIN'] }]);
-    const result = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/valid-serverless-regex.yaml'),
-      fileContent: undefined,
-      fileType: 'MAIN',
-      tsConfigs: [],
-    });
-    const {
-      issues: [issue],
-    } = result;
-    expect(issue).toEqual(
-      expect.objectContaining({
-        line: 7,
-        column: 41,
-        endLine: 7,
-        endColumn: 44,
-      }),
-    );
-  });
-
-  it('should not return issues outside of the JS snippets', async () => {
-    initLinter([
-      { key: 'no-trailing-spaces', configurations: [], fileTypeTarget: ['MAIN'] },
-      { key: 'file-header', configurations: [{ headerFormat: '' }], fileTypeTarget: ['MAIN'] },
-    ]);
-    const { issues } = await analyzeYaml({
-      filePath: join(__dirname, './fixtures/yaml/trailing-spaces.yaml'),
-      fileContent: undefined,
-      fileType: 'MAIN',
-      tsConfigs: [],
-    });
-    expect(issues).toHaveLength(0);
-  });
-
-  it('should throw without init-linter', () => {
-    try {
-      jest.resetModules();
-      const analyzer = require('../src/analyzer');
-      analyzer.analyzeYaml({
-        filePath: '/fake/path',
-        fileContent: 'fakeCode();',
-      });
-      fail('this test should throw error');
-    } catch (e) {
-      expect(e.message).toBe('Linter is undefined. Did you call /init-linter?');
-    }
   });
 });
