+++ eslint-bridge/tests/parser.test.ts
@@ -20,15 +20,12 @@
 import {
   buildParsingOptions,
   buildSourceCode,
-  buildSourceCodesFromYaml,
   ParseExceptionCode,
   parseExceptionCodeOf,
-  parseYaml,
 } from 'parser';
 import { SourceCode } from 'eslint';
 import { ParsingError, ProgramBasedAnalysisInput } from 'analyzer';
 import { visit } from '../src/utils';
-import * as estree from 'estree';
 import * as path from 'path';
 import * as fs from 'fs';
 import { setContext } from 'context';
@@ -386,290 +383,6 @@ describe('program based analysis', () => {
     const parsingOptions = buildParsingOptions(analysisInput);
     expect(parsingOptions.programs).toHaveLength(1);
     expect(parsingOptions.project).toBeUndefined();
-  });
-});
-
-describe('parse YAML Files', () => {
-  const YAML_LAMBDA_FILE_PATH = join(__dirname, './fixtures/yaml/valid-lambda.yaml');
-  const YAML_SERVERLESS_FILE_PATH = join(__dirname, './fixtures/yaml/valid-serverless.yaml');
-  const INVALID_YAML_FILE_PATH = join(__dirname, './fixtures/yaml/invalid-yaml.yaml');
-  const INVALID_PLAIN_INLINE_JS_FILE_PATH = join(
-    __dirname,
-    './fixtures/yaml/invalid-plain-inline-js.yaml',
-  );
-  const INVALID_BLOCK_INLINE_JS_FILE_PATH = join(
-    __dirname,
-    './fixtures/yaml/invalid-block-inline-js.yaml',
-  );
-  const PLAIN_FORMAT_FILE_PATH = join(__dirname, './fixtures/yaml/flow-plain.yaml');
-  const BLOCK_FOLDED_FORMAT_FILE_PATH = join(__dirname, './fixtures/yaml/block-folded.yaml');
-  const BLOCK_LITERAL_FORMAT_FILE_PATH = join(__dirname, './fixtures/yaml/block-literal.yaml');
-  const UNSUPPORTED_FORMAT_FILE_PATH = join(__dirname, './fixtures/yaml/unsupported-format.yaml');
-
-  it('should parse YAML syntax', () => {
-    const parsed = parseYaml(YAML_LAMBDA_FILE_PATH);
-    expect(parsed).toBeDefined();
-    expect(parsed).toHaveLength(1);
-    expect(parsed[0]).toEqual(
-      expect.objectContaining({
-        code: `if (foo()) bar(); else bar();`,
-        line: 8,
-        column: 18,
-        offset: 177,
-      }),
-    );
-  });
-
-  it('should build source code from YAML lambda file', () => {
-    const sourceCodes = buildSourceCodesFromYaml(YAML_LAMBDA_FILE_PATH);
-    expect(sourceCodes).toHaveLength(1);
-    expect(sourceCodes[0]).toBeInstanceOf(SourceCode);
-    expect(sourceCodes[0].ast.loc.start).toEqual({ line: 8, column: 17 });
-  });
-
-  it('should build source code from YAML serverless file', () => {
-    const sourceCodes = buildSourceCodesFromYaml(YAML_SERVERLESS_FILE_PATH);
-    expect(sourceCodes).toHaveLength(1);
-    expect(sourceCodes[0]).toBeInstanceOf(SourceCode);
-    expect(sourceCodes[0].ast.loc.start).toEqual({ line: 7, column: 18 });
-  });
-
-  it('should handle YAML parsing errors', () => {
-    const parsingError = buildSourceCodesFromYaml(INVALID_YAML_FILE_PATH) as ParsingError;
-    expect(parsingError).toEqual({
-      code: ParseExceptionCode.Parsing,
-      line: 2,
-      message: 'Map keys must be unique',
-    });
-  });
-
-  it('should return a parsing error when parsing invalid plain inline JS', () => {
-    const parsingError = buildSourceCodesFromYaml(
-      INVALID_PLAIN_INLINE_JS_FILE_PATH,
-    ) as ParsingError;
-    expect(parsingError).toEqual({
-      code: ParseExceptionCode.Parsing,
-      line: 7,
-      message: `Unexpected token ','. (7:22)`,
-    });
-  });
-
-  it('should return a parsing error when parsing invalid block inline JS', () => {
-    const parsingError = buildSourceCodesFromYaml(
-      INVALID_BLOCK_INLINE_JS_FILE_PATH,
-    ) as ParsingError;
-    expect(parsingError).toEqual({
-      code: ParseExceptionCode.Parsing,
-      line: 8,
-      message: `Unexpected token ','. (8:15)`,
-    });
-  });
-
-  it('it should not build a source code for an unsupported format', () => {
-    const sourceCodes = buildSourceCodesFromYaml(UNSUPPORTED_FORMAT_FILE_PATH);
-    expect(sourceCodes).toHaveLength(0);
-  });
-
-  it('should fix plain-based format locations', () => {
-    const [{ ast }] = buildSourceCodesFromYaml(PLAIN_FORMAT_FILE_PATH) as SourceCode[];
-
-    const {
-      body: [ifStmt],
-    } = ast;
-    expect(ifStmt.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 7,
-          column: 18,
-        },
-        end: {
-          line: 7,
-          column: 67,
-        },
-      }),
-    );
-    expect(ifStmt.range).toEqual([170, 219]);
-
-    const { alternate } = ifStmt as estree.IfStatement;
-    expect(alternate.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 7,
-          column: 57,
-        },
-        end: {
-          line: 7,
-          column: 67,
-        },
-      }),
-    );
-    expect(alternate.range).toEqual([209, 219]);
-
-    const {
-      comments: [comment],
-    } = ast;
-    expect(comment.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 7,
-          column: 38,
-        },
-        end: {
-          line: 7,
-          column: 49,
-        },
-      }),
-    );
-    expect(comment.range).toEqual([190, 201]);
-
-    const elseToken = ast.tokens.find(token => token.value === 'else');
-    expect(elseToken.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 7,
-          column: 52,
-        },
-        end: {
-          line: 7,
-          column: 56,
-        },
-      }),
-    );
-    expect(elseToken.range).toEqual([204, 208]);
-  });
-
-  it('should fix block-folded-based format locations', () => {
-    const [{ ast }] = buildSourceCodesFromYaml(BLOCK_FOLDED_FORMAT_FILE_PATH) as SourceCode[];
-    const {
-      body: [ifStmt],
-    } = ast;
-    expect(ifStmt.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 8,
-          column: 8,
-        },
-        end: {
-          line: 12,
-          column: 9,
-        },
-      }),
-    );
-    expect(ifStmt.range).toEqual([180, 265]);
-
-    const { alternate } = ifStmt as estree.IfStatement;
-    expect(alternate.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 10,
-          column: 15,
-        },
-        end: {
-          line: 12,
-          column: 9,
-        },
-      }),
-    );
-    expect(alternate.range).toEqual([237, 265]);
-
-    const {
-      comments: [comment],
-    } = ast;
-    expect(comment.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 9,
-          column: 17,
-        },
-        end: {
-          line: 9,
-          column: 28,
-        },
-      }),
-    );
-    expect(comment.range).toEqual([210, 221]);
-
-    const elseToken = ast.tokens.find(token => token.value === 'else');
-    expect(elseToken.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 10,
-          column: 10,
-        },
-        end: {
-          line: 10,
-          column: 14,
-        },
-      }),
-    );
-    expect(elseToken.range).toEqual([232, 236]);
-  });
-
-  it('should fix block-literal-based format locations', () => {
-    const [{ ast }] = buildSourceCodesFromYaml(BLOCK_LITERAL_FORMAT_FILE_PATH) as SourceCode[];
-    const {
-      body: [ifStmt],
-    } = ast;
-    expect(ifStmt.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 8,
-          column: 8,
-        },
-        end: {
-          line: 12,
-          column: 9,
-        },
-      }),
-    );
-    expect(ifStmt.range).toEqual([180, 265]);
-
-    const { alternate } = ifStmt as estree.IfStatement;
-    expect(alternate.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 10,
-          column: 15,
-        },
-        end: {
-          line: 12,
-          column: 9,
-        },
-      }),
-    );
-    expect(alternate.range).toEqual([237, 265]);
-
-    const {
-      comments: [comment],
-    } = ast;
-    expect(comment.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 9,
-          column: 17,
-        },
-        end: {
-          line: 9,
-          column: 28,
-        },
-      }),
-    );
-    expect(comment.range).toEqual([210, 221]);
-
-    const elseToken = ast.tokens.find(token => token.value === 'else');
-    expect(elseToken.loc).toEqual(
-      expect.objectContaining({
-        start: {
-          line: 10,
-          column: 10,
-        },
-        end: {
-          line: 10,
-          column: 14,
-        },
-      }),
-    );
-    expect(elseToken.range).toEqual([232, 236]);
   });
 });
